Fix infinite loop when reading an ambiguous chunk
When a chunk is completely ambiguous, it would pass the empty to the callback,
and unread the whole block. The next iteration would to the same.

Instead, this additionally prefetches the next chunk. If the read fails, the
block could be sent as application data, but valid multipart/form-data messages
always end with the boundary string.